requirements.txt
streamlit
segno
Pillow